Estoque passa a aceitar quantidade fracionária (produto vendido por peso/metro)
A coluna era inteira, truncando estoque fracionário vindo do Link Pro
(produto vendido por peso ou metro). Agora é decimal(12,3) — cobre venda,
ajuste manual e sincronização do agente sem arredondar.

+++ app/Http/Controllers/Api/ProdutoController.php
@@ -88,7 +88,7 @@ class ProdutoController extends Controller
     {
         $data = $request->validate([
             'loja_id' => ['required', 'exists:lojas,id'],
-            'quantidade' => ['required', 'integer'],
+            'quantidade' => ['required', 'numeric'],
         ]);
 
         $estoque = ProdutoEstoque::updateOrCreate(
